+++ src/main/java/audaki/cart_engine/AceGameRules.java
@@ -5,11 +5,20 @@ import net.fabricmc.fabric.api.gamerule.v1.GameRuleRegistry;
 import net.minecraft.world.level.GameRules;
 
 public class AceGameRules {
-    public static GameRules.Key<GameRules.IntegerValue> ACE_CART_SPEED;
+    // All speeds in blocks per second
+    public static GameRules.Key<GameRules.IntegerValue> MINECART_MAX_SPEED_PLAYER_RIDER;
+    public static GameRules.Key<GameRules.IntegerValue> MINECART_MAX_SPEED_OTHER_RIDER;
+    public static GameRules.Key<GameRules.IntegerValue> MINECART_MAX_SPEED_EMPTY_RIDER;
 
     public static void register() {
-        ACE_CART_SPEED = GameRuleRegistry.register("aceCartSpeed",
+        MINECART_MAX_SPEED_PLAYER_RIDER = GameRuleRegistry.register("minecartMaxSpeedPlayerRider",
                 GameRules.Category.PLAYER,
                 GameRuleFactory.createIntRule(20));
+        MINECART_MAX_SPEED_OTHER_RIDER = GameRuleRegistry.register("minecartMaxSpeedOtherRider",
+                GameRules.Category.PLAYER,
+                GameRuleFactory.createIntRule(0));
+        MINECART_MAX_SPEED_EMPTY_RIDER = GameRuleRegistry.register("minecartMaxSpeedEmptyRider",
+                GameRules.Category.PLAYER,
+                GameRuleFactory.createIntRule(0));
     }
 }