+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.0] - 2024-12-15
+- Add more configurable speeds via /gamerule
+  - minecartMaxSpeed: Generic max speed for all carts unless overridden
+  - minecartMaxSpeedPlayerRider: If not 0, overrides speed only for player ridden carts
+  - minecartMaxSpeedOtherRider: If not 0, overrides speed only for non-player ridden carts
+  - minecartMaxSpeedEmptyRider: If not 0, overrides speed only for rideable carts that are empty
+- (Removed gamerule aceCartSpeed)
+
 ## [4.0.0] - 2024-12-15
 - Rebase mod on improved minecarts experiment
   - The experimental code will be activated for all carts